# backend/src/main/java/by/gusto/auth/dto/TemporaryPasswordResponse.java
package by.gusto.auth.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class TemporaryPasswordResponse {

    private UUID userId;
    private String temporaryPassword;
}