# shared/openapi.yaml
openapi: 3.1.0
info:
  title: GUSTO B2B API
  version: 0.2.0
  description: |
    Контракт API B2B-портала «Густо». Заполняется по мере сессий (чек-лист, Часть 4):
    каждая сессия, меняющая эндпоинт, обновляет этот файл в том же PR.

    **Envelope.** Все ответы: `{data, meta, error}`. Успех — `data` заполнен, `error: null`.
    Ошибка — `data: null`, `error: {code, message, details?}`.

    **Каталог кодов ошибок** (`error.code`) фиксируется в `ErrorCode` и дальше только
    расширяется — существующие коды не переименовываются и не меняют смысл.

    **Пагинация** списков: query `page` (0-based, как в Spring Data), `size`;
    ответ — `meta: {page, size, total}` (total — всего элементов).

    **Аутентификация.** Access-токен (JWT, 15 мин) — заголовок `Authorization: Bearer`.
    Refresh-токен (7 дней, rotation + reuse detection) — httpOnly-кука
    `refresh_token` (Secure, SameSite=Strict), выдаётся при login/refresh,
    используется только эндпоинтом `/auth/refresh` и инвалидируется `/auth/logout`.

    **Идемпотентность.** Мутирующие `POST` (заказы, заявки) принимают заголовок
    `Idempotency-Key` (см. `components.parameters.IdempotencyKey`): повтор с тем же
    ключом в течение 24 ч возвращает тот же результат.

    **Health-эндпоинты** `/healthz`, `/readyz` живут вне префикса `/api/v1`
    (per-path `servers`).
servers:
  - url: /api/v1
...
paths:
...
  /admin/users:
    get:
      tags: [admin]
      summary: Список пользователей (только ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - name: role
          in: query
          schema: { $ref: "#/components/schemas/Role" }
        - name: companyId
          in: query
          schema: { type: string, format: uuid }
        - name: managerId
          in: query
          schema: { type: string, format: uuid }
        - name: search
          in: query
          schema: { type: string }
          description: Поиск по email, ФИО, телефону
        - name: isActive
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Список пользователей
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      summary: Создание пользователя ADMIN/ACCOUNTANT/MANAGER/CUSTOMER_LEGAL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: Пользователь создан; пароль выдаётся админу отдельно
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID пользователя
        schema: { type: string, format: uuid }
    get:
      tags: [admin]
      summary: Карточка пользователя
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Данные пользователя
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [admin]
      summary: Редактирование пользователя
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        "200":
          description: Пользователь обновлён
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [admin]
      summary: Мягкое удаление пользователя
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Пользователь удалён
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/users/{id}/reset-password:
    parameters:
      - name: id
        in: path
        required: true
        description: ID пользователя
        schema: { type: string, format: uuid }
    post:
      tags: [admin]
      summary: Сброс пароля админом (временный пароль, показывается 1 раз)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Временный пароль
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TemporaryPasswordResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/companies:
    get:
      tags: [admin]
      summary: Список компаний
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/CompanyStatus" }
        - name: managerId
          in: query
          schema: { type: string, format: uuid }
        - name: search
          in: query
          schema: { type: string }
          description: Поиск по названию, сокращённому названию, УНП
      responses:
        "200":
          description: Список компаний
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompanyListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      summary: Создание компании
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCompanyRequest" }
      responses:
        "201":
          description: Компания создана
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompanyResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/companies/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID компании
        schema: { type: string, format: uuid }
    get:
      tags: [admin]
      summary: Карточка компании
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Данные компании
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompanyResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [admin]
      summary: Редактирование компании
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCompanyRequest" }
      responses:
        "200":
          description: Компания обновлена
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompanyResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [admin]
      summary: Деактивация компании (status = INACTIVE)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Компания деактивирована
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-токен из /auth/login или /auth/refresh (15 мин)
  parameters:
...
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
      description: Номер страницы, 0-based
    SizeParam:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    Unauthorized:
      description: Не аутентифицирован / неверные креды / истёкший токен
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            invalidCredentials:
              value: { data: null, meta: { }, error: { code: AUTH_INVALID_CREDENTIALS, message: "Неверный email или пароль" } }
    ValidationFailed:
      description: Ошибка валидации полей запроса
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            invalidFields:
              value: { data: null, meta: { }, error: { code: VALIDATION_FAILED, message: "Ошибка валидации", details: { "email": "некорректный email" } } }
    Forbidden:
      description: Доступ запрещён (недостаточно прав или чужой ресурс)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            forbidden:
              value: { data: null, meta: { }, error: { code: ACCESS_DENIED, message: "Доступ запрещён" } }
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            notFound:
              value: { data: null, meta: { }, error: { code: NOT_FOUND, message: "Не найдено" } }
    Conflict:
      description: Конфликт уникальности (email/УНП уже заняты)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            duplicate:
              value: { data: null, meta: { }, error: { code: CONFLICT, message: "Пользователь с таким email уже существует" } }
...
  schemas:
...
    Envelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          nullable: true
          description: Полезная нагрузка; null при ошибке
        meta:
          $ref: "#/components/schemas/Meta"
        error:
          nullable: true
          $ref: "#/components/schemas/Error"
    Meta:
      type: object
      description: Для списков — пагинация; иначе пустой объект
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0, description: "Всего элементов" }
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: Человекочитаемое сообщение (ru)
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: "Опционально: поля формы → текст ошибки, и т.п."
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data: { nullable: true, enum: [null] }
            error:
              $ref: "#/components/schemas/Error"

    # Каталог кодов ошибок error.code. Только расширяется; не переименовывается.
    ErrorCode:
      type: string
      enum:
        # auth
        - AUTH_INVALID_CREDENTIALS # неверный email/пароль
        - AUTH_UNAUTHORIZED        # запрос без/с недействительным access-токеном
        - AUTH_2FA_REQUIRED        # login без totpCode при включённой 2FA
        - AUTH_2FA_INVALID         # неверный код TOTP
        - AUTH_TOKEN_EXPIRED       # access-токен истёк
        - AUTH_TOKEN_INVALID       # подпись/формат токена не валидны
        - AUTH_REFRESH_INVALID     # refresh-токен недействителен/отсутствует
        - AUTH_REFRESH_REUSED      # повтор использованного refresh → ревок цепочки
        - AUTH_RESET_TOKEN_INVALID # токен восстановления просрочен/неверен
        # доступ
        - ACCESS_DENIED            # 403 по роли/ownership
        - NOT_FOUND                # 404 сущность не найдена
        - CONFLICT                 # 409 конфликт уникальности
        - RATE_LIMITED             # 429 превышен лимит
        - IDEMPOTENCY_CONFLICT     # тот же ключ, другое тело запроса
        # валидация
        - VALIDATION_FAILED        # ошибки полей запроса
        # склад
        - STOCK_INSUFFICIENT       # недостаточно остатка для заказа/резерва
        - STOCK_DOCUMENT_INVALID   # неверный переход статуса складского документа
        # общие
        - INTERNAL                 # непредвиденная ошибка сервера

    # ----- Auth ------------------------------------------------------------------
    Role:
      type: string
      enum: [ADMIN, ACCOUNTANT, MANAGER, CUSTOMER_LEGAL, CUSTOMER_INDIVIDUAL]
    User:
      type: object
      required: [id, email, fullName, role, isActive, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        fullName: { type: string }
        phone: { type: string, nullable: true }
        role: { $ref: "#/components/schemas/Role" }
        companyId: { type: string, format: uuid, nullable: true, description: "Заполнен у CUSTOMER_LEGAL" }
        isActive: { type: boolean }
        totpEnabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    UserResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data: { $ref: "#/components/schemas/User" }
    UserListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/User" }
    CompanyStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    Company:
      type: object
      required: [id, name, status]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        shortName: { type: string, nullable: true }
        unp: { type: string, nullable: true, pattern: "^\\d{9}|\\d{10}$", description: "УНП РБ: 9 или 10 цифр" }
        legalAddress: { type: string, nullable: true }
        actualAddress: { type: string, nullable: true }
        bankAccount: { type: string, nullable: true }
        bankName: { type: string, nullable: true }
        bankBic: { type: string, nullable: true }
        contactPhone: { type: string, nullable: true }
        contactEmail: { type: string, format: email, nullable: true }
        status: { $ref: "#/components/schemas/CompanyStatus" }
        managerId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CompanyResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data: { $ref: "#/components/schemas/Company" }
    CompanyListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Company" }
...
    CreateUserRequest:
      type: object
      required: [email, fullName, role]
      properties:
        email: { type: string, format: email }
        fullName: { type: string, minLength: 2, maxLength: 255 }
        phone: { type: string, nullable: true }
        role: { $ref: "#/components/schemas/Role" }
        companyId: { type: string, format: uuid, nullable: true, description: "Обязательна для CUSTOMER_LEGAL" }
    UpdateUserRequest:
      type: object
      properties:
        fullName: { type: string, minLength: 2, maxLength: 255 }
        phone: { type: string, nullable: true }
        role: { $ref: "#/components/schemas/Role" }
        companyId: { type: string, format: uuid, nullable: true }
        isActive: { type: boolean }
    TemporaryPasswordResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data:
              type: object
              required: [userId, temporaryPassword]
              properties:
                userId: { type: string, format: uuid }
                temporaryPassword: { type: string, description: "Показывается один раз" }
    CreateCompanyRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 255 }
        shortName: { type: string, maxLength: 255, nullable: true }
        unp: { type: string, pattern: "^\\d{9}|\\d{10}$", nullable: true, description: "УНП РБ: 9 или 10 цифр" }
        legalAddress: { type: string, maxLength: 500, nullable: true }
        actualAddress: { type: string, maxLength: 500, nullable: true }
        bankAccount: { type: string, maxLength: 50, nullable: true }
        bankName: { type: string, maxLength: 255, nullable: true }
        bankBic: { type: string, maxLength: 20, nullable: true }
        contactPhone: { type: string, maxLength: 50, nullable: true }
        contactEmail: { type: string, format: email, maxLength: 255, nullable: true }
        managerId: { type: string, format: uuid, nullable: true }
        status: { $ref: "#/components/schemas/CompanyStatus" }
    UpdateCompanyRequest:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        shortName: { type: string, maxLength: 255, nullable: true }
        unp: { type: string, pattern: "^\\d{9}|\\d{10}$", nullable: true }
        legalAddress: { type: string, maxLength: 500, nullable: true }
        actualAddress: { type: string, maxLength: 500, nullable: true }
        bankAccount: { type: string, maxLength: 50, nullable: true }
        bankName: { type: string, maxLength: 255, nullable: true }
        bankBic: { type: string, maxLength: 20, nullable: true }
        contactPhone: { type: string, maxLength: 50, nullable: true }
        contactEmail: { type: string, format: email, maxLength: 255, nullable: true }
        managerId: { type: string, format: uuid, nullable: true }
        status: { $ref: "#/components/schemas/CompanyStatus" }
...